refactor(defense): replace commission members with flexible assignment system
- Remove CommissionMember entity, replace with generic Assignment model
- Remove Expert entity and ExpertRepository (use Assignment instead)
- Update Commission entity to use Assignments navigation property
- Refactor all commission commands to work with AssignmentId instead of MemberId
- Update Grade entity to reference Assignment instead of CommissionMember
- Remove CommissionMemberConfiguration and ExpertConfiguration
- Update CommissionRepository methods to include Assignments
- Modify all defense-related commands and queries for new model
- Update CommissionsController API endpoints (memberId → assignmentId)

+++ src/AWM.Service.Application/Features/Defense/Commissions/Commands/AddCommissionMember/AddCommissionMemberCommand.cs
@@ -6,7 +6,7 @@ using MediatR;
 /// <summary>
 /// Command to add a member to a commission.
 /// </summary>
-public sealed record AddCommissionMemberCommand : IRequest<Result<int>>
+public sealed record AddCommissionMemberCommand : IRequest<Result<long>>
 {
     /// <summary>
     /// Commission ID to add the member to.
